manim-service/requirements.txt
flask==2.0.1
gunicorn==20.1.0
werkzeug==2.0.3
manim==0.17.3
supabase==1.0.3
python-dotenv==1.0.0
requests==2.31.0